% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcFit.R
\name{calculateBinProbabilities}
\alias{calculateBinProbabilities}
\title{calculateBinProbabilities}
\usage{
calculateBinProbabilities(resOb)
}
\arguments{
\item{resOb}{Observed data (see dmcObservedData)}
}
\value{
resOb Observed data with additional $probSubject/$prob table
}
\description{
Calculate bin probabilities in observed data
}
\examples{
# Example 1:
resOb <- flankerData
resOb <- calculateBinProbabilities(resOb)
resOb$prob
}
