% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCostValue.R
\name{calculateCostValue}
\alias{calculateCostValue}
\title{calculateCostValue}
\usage{
calculateCostValue(resTh, resOb)
}
\arguments{
\item{resTh}{list containing caf values for comp/incomp conditions (n*2*3) and
delta values for comp/incomp conditions (n*2*5). See output from dmcSim.}

\item{resOb}{list containing caf values for comp/incomp conditions (n*2*3) and
delta values for comp/incomp conditions (n*5). See output from dmcSim.}
}
\value{
cost
}
\description{
Calculate cost value (fit) as combination of RT and error rate
}
\examples{
library(DMCfun)

# Example 1:
resTh <- dmcSim()
resOb <- dmcSim()
cost  <- calculateCostValue(resTh, resOb)

# Example 2:
resTh <- dmcSim()
resOb <- dmcSim(tau = 150)
cost  <- calculateCostValue(resTh, resOb)

}
