% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcFit.R
\name{calculateCostValueCS}
\alias{calculateCostValueCS}
\title{calculateCostValueCS}
\usage{
calculateCostValueCS(resTh, resOb)
}
\arguments{
\item{resTh}{list containing simulation $sim values (output from dmcSim) for rts_comp, rts_incomp,
errs_comp, errs_incomp}

\item{resOb}{list containing raw observed data (see dmcObservedData with keepRaw = TRUE)}
}
\value{
cost value (CS)
}
\description{
Calculate cost value (fit) using chi-square (CS) from correct and incorrect RT data.
}
\examples{
# Example 1:
resTh <- dmcSim()
resOb <- flankerData
resOb <- calculateBinProbabilities(resOb)
cost  <- calculateCostValueCS(resTh, resOb)
}
