% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcFit.R
\name{calculateCostValueGS}
\alias{calculateCostValueGS}
\title{calculateCostValueGS: Calculate likelihood-ratio chi-square statistic (GS) statistic from reaction times
for both correct and incorrect trials}
\usage{
calculateCostValueGS(resTh, resOb)
}
\arguments{
\item{resTh}{list containing simulation $sim values (output from dmcSim) for rts_comp, rts_incomp,
errs_comp, errs_incomp}

\item{resOb}{list containing raw observed data (see dmcObservedData with keepRaw = TRUE}
}
\value{
cost value (GS)
}
\description{
Calculate cost value (fit) from correct and incorrect RT data.
}
\examples{
# Example 1:
resTh <- dmcSim()
resOb <- flankerData
resOb <- calculateBinProbabilities(resOb)
cost <- calculateCostValueGS(resTh, resOb)
}
