% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDF.R
\name{createDF}
\alias{createDF}
\title{createDF}
\usage{
createDF(
  nVP = 20,
  nTrl = 50,
  design = list(A = c("A1", "A2"), B = c("B1", "B2"))
)
}
\arguments{
\item{nVP}{Number of participants}

\item{nTrl}{Number of trials per factor/level for each participant}

\item{design}{Factors and levels}
}
\value{
dataframe
}
\description{
Create dataframe (see also addDataDF)
}
\examples{
# Example 1
dat <- createDF()

# Example 2
dat <- createDF(nVP = 50, nTrl = 50, design = list("Comp" = c("comp", "incomp")))

# Example 3
dat <- createDF(nVP = 50, nTrl = 50, design = list("Comp" = c("comp", "incomp"),
                                                   "Side" = c("left", "right", "middle")))

}
