% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcData.R
\name{dmcCombineObservedData}
\alias{dmcCombineObservedData}
\title{dmcCombineObservedData}
\usage{
dmcCombineObservedData(...)
}
\arguments{
\item{...}{Any number of outputs from dmcObservedData}
}
\value{
dmcCombineObservedData returns a list of objects of class "dmcob"
}
\description{
Combine observed datasets
}
\examples{
# Example 1
dat <- dmcCombineObservedData(flankerData, simonData)  # combine flanker/simon data
plot(dat, figType = "delta", xlimDelta = c(200, 700), ylimDelta = c(-20, 80),
     cols = c("black", "darkgrey"), legend.parameters = list(x=200, y=80,
     legend = c("Flanker Task", "Simon Task")))

}
