% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMCfun.R
\docType{package}
\name{DMCfun-package}
\alias{DMCfun}
\alias{DMCfun-package}
\title{DMCfun}
\description{
DMC model simulation detailed in Ulrich, R., Schroeter, H., Leuthold, H., & Birngruber, T. (2015). Automatic and controlled stimulus processing in conflict tasks: Superimposed diffusion processes and delta functions. Cognitive Psychology, 78, 148-174.
}
\author{
\strong{Maintainer}: Mackenzie G. Ian \email{ian.mackenzie@uni-tuebingen.de}

Authors:
\itemize{
  \item Leuthold Hartmut \email{hartmut.leuthold@uni-tuebingen.de}
  \item Ulrich Rolf \email{rolf.ulrich@uni-tuebingen.de}
}

}
