% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcObservedData.R
\name{dmcObservedData}
\alias{dmcObservedData}
\title{dmcObservedData}
\usage{
dmcObservedData(
  dat,
  stepCAF = 20,
  stepDelta = 5,
  outlier = c(200, 1200),
  quantileType = 5,
  columns = c("VP", "Comp", "RT", "Error"),
  compCoding = c("comp", "incomp"),
  errorCoding = c(0, 1)
)
}
\arguments{
\item{dat}{Text file containing the observed data or R dataframe (see createDF/addDataDF)}

\item{stepCAF}{Step size for the CAF bins. For example, a step size of 20 would result
in 5 CAF bins centered on 10, 30, 50, 70, and 90\%.}

\item{stepDelta}{Step size for the Delta bins. For example, a step size of 5 would result
in 19 CAF bins positioned at 5, 10, 15, ... 85, 90, 95\%.}

\item{outlier}{Outlier limits e.g., c(200, 1200)}

\item{quantileType}{Argument (1-9) from R function quantile specifying the algorithm (?quantile)}

\item{columns}{Name of required columns DEFAULT = c("VP", "Comp", "RT", "Error")}

\item{compCoding}{Coding for compatibility DEFAULT = c("comp", "incomp")}

\item{errorCoding}{Coding for errors DEFAULT = c(0, 1))}
}
\value{
DataFrame
}
\description{
Basic example analysis script to create data object required
for observed data. Example raw *.txt files are flankerData.txt and simonData.txt.
}
\examples{
\dontrun{
library(psychReport)
requiredPackages(c("DMCfun", "ez", "dplyr"))

# Example 1
# datOb <- dmcObservedData("flankerData.txt")

# Example 2
# datOb <- dmcObservedData("simonData.txt")

# Example 3 (Basic behavioural analysis from Ulrich et al. 2015)
flankerData$summaryVP <- tibble::add_column(Task = "flanker", flankerData$summaryVP, .before = 2)
simonData$summaryVP   <- tibble::add_column(Task = "simon", simonData$summaryVP, .before = 2)
datAgg <- rbind(rbind(flankerData$summaryVP, simonData$summaryVP))

datAgg$VP   <- factor(datAgg$VP)
datAgg$Task <- factor(datAgg$Task)
datAgg$Comp <- factor(datAgg$Comp)

aovErr <- ezANOVA(datAgg, dv = .(perErr), wid = .(VP), within = .(Comp, Task),
                  detailed = TRUE, return_aov = TRUE)
aovErr <- aovTable(aovErr)
aovDispMeans(aovErr)

aovRt <- ezANOVA(datAgg, dv = .(rtCor), wid = .(VP), within = .(Comp, Task),
                 detailed = TRUE, return_aov = TRUE)
aovRt <- aovTable(aovRt)
aovDispMeans(aovRt)

# Delta
flankerData$deltaVP <- tibble::add_column(Task = "flanker", flankerData$deltaVP, .before = 2)
simonData$deltaVP   <- tibble::add_column(Task = "simon", simonData$deltaVP, .before = 2)
datDelta             <- rbind(rbind(flankerData$deltaVP, simonData$deltaVP))

datDelta <- datDelta \%>\%
  dplyr::select(-c("meanBin", "meanEffect")) \%>\%
  tidyr::gather(Comp, rt, -c("VP", "Task", "bin"))

datDelta$VP   <- factor(datDelta$VP)
datDelta$Task <- factor(datDelta$Task)
datDelta$bin  <- factor(datDelta$bin)
datDelta$Comp <- factor(datDelta$Comp)

aovDelta <- ezANOVA(datDelta, dv = .(rt), wid = .(VP), within = .(Task, Comp, bin),
                    detailed = TRUE, return_aov = TRUE)
aovDelta <- aovTable(aovDelta)

# CAF
flankerData$cafVP <- tibble::add_column(Task = "flanker", flankerData$cafVP, .before = 2)
simonData$cafVP   <- tibble::add_column(Task = "simon", simonData$cafVP, .before = 2)
datCaf             <- rbind(rbind(flankerData$cafVP, simonData$cafVP))

datCaf$VP   <- factor(datCaf$VP)
datCaf$Task <- factor(datCaf$Task)
datCaf$bin  <- factor(datCaf$bin)
datCaf$Comp <- factor(datCaf$Comp)

aovCaf <- ezANOVA(datCaf, dv = .(accPer), wid = .(VP), within = .(Task, Comp, bin),
                  detailed = TRUE, return_aov = TRUE)
aovCaf <- aovTable(aovCaf)
aovDispMeans(aovCaf)

}
# Example 4
dat <- createDF(nVP = 50, nTrl = 500,
                design = list("Comp" = c("comp", "incomp")))
dat <- addDataDF(dat,
                 RT = list(list(c("Comp:comp"), vals = c(500, 75, 120)),
                           list(c("Comp:incomp"), vals = c(530, 75, 100))),
                 Error = list(list(c("Comp:comp"), vals = c(3, 2, 2, 1, 1)),
                            list(c("Comp:incomp"), vals = c(15, 3, 2, 1, 1))))
datOb <- dmcObservedData(dat, stepCAF = 10, stepDelta = 10)
plot(datOb)
plot(datOb, VP = 1)

}
