% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcObservedDataRaw.R
\name{dmcObservedDataRaw}
\alias{dmcObservedDataRaw}
\title{dmcObservedDataRaw}
\usage{
dmcObservedDataRaw(
  datafile,
  stepCAF = 20,
  stepDelta = 5,
  outlier = c(200, 1200),
  quantileType = 5,
  datDir = NULL,
  saveData = FALSE
)
}
\arguments{
\item{datafile}{Text file containing the observed data}

\item{stepCAF}{Number of bins for the conditional accuracy function (CAF)}

\item{stepDelta}{Number of bins fore the distributional analysis}

\item{outlier}{Outlier limits e.g., c(200, 1200)}

\item{quantileType}{1}

\item{datDir}{Data directory}

\item{saveData}{TRUE/FALSE}
}
\value{
DataFrame
}
\description{
Basic example analysis script to create data object required
for observed data. Example raw *.txt files are included.
}
\examples{
\dontrun{
# Example 1
datOb <- dmcObservedDataRaw("flankerData1.txt", saveData = TRUE)

# Example 2
datOb <- dmcObservedDataRaw("flankerData2.txt", saveData = TRUE)

# Example 3
datOb <- dmcObservedDataRaw("simonData1.txt", saveData = TRUE)

# Example 4
datOb <- dmcObservedDataRaw("simonData2.txt", saveData = TRUE)

}
}
