% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcSim.R
\name{dmcSimApp}
\alias{dmcSimApp}
\title{dmcSimApp}
\usage{
dmcSimApp()
}
\value{
Shiny app

Shiny App

A shiny app allowing interactive exploration of DMC parameters
}
\description{
Shiny interface to function dmcSim
}
\examples{
\donttest{
dmcSimApp()
}

}
