% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcSimR.R
\name{dmcSimR}
\alias{dmcSimR}
\title{dmcSimR}
\usage{
dmcSimR(amp = 20, tau = 30, aaShape = 2, mu = 0.5, sigma = 4,
  bnds = 75, resMean = 300, resSD = 30, nTrl = 1e+05,
  tmax = 1000, varSP = FALSE, spShape = 3, spLim = c(-75, 75),
  varDR = FALSE, drShape = 3, drLim = c(0.1, 0.7), nTrlData = 5,
  stepDelta = 5, stepCAF = 20, setSeed = FALSE)
}
\arguments{
\item{amp}{amplitude of automatic activation}

\item{tau}{time to peak automatic activation}

\item{aaShape}{shape parameter of automatic activation}

\item{mu}{drift rate of controlled processes}

\item{sigma}{diffusion constant}

\item{bnds}{+- response barrier}

\item{resMean}{mean of non-decisional component}

\item{resSD}{standard deviation of non-decisional component}

\item{nTrl}{number of trials}

\item{tmax}{number of time points per trial}

\item{varSP}{TRUE/FALSE variable starting point}

\item{spShape}{shape parameter of drift rate}

\item{spLim}{limit range of distribution of starting point}

\item{varDR}{TRUE/FALSE variable drift rate}

\item{drShape}{shape parameter of drift rate}

\item{drLim}{limit range of distribution of drift rate}

\item{nTrlData}{Number of trials to plot}

\item{stepDelta}{Step size for the Delta bins. For example, a step size of 5 would result
in 19 CAF bins positioned at 5, 10, 15, ... 85, 90, 95\%.}

\item{stepCAF}{Step size for the CAF bins. For example, a step size of 20 would result
in 5 CAF bins centered on 10, 30, 50, 70, and 90\%.}

\item{setSeed}{TRUE/FALSE If TRUE, set.seed(1) is called}
}
\value{
dmcsim

The function returns a list with the relevant results from the simulation. The list
is accessed with obj$name with name being one of the following:
\item{obj$summary}{summary}
\item{obj$delta}{delta}
\item{obj$caf}{caf}
\item{obj$sim}{sim}
\item{obj$trials}{trials}
\item{obj$prms}{prms}
}
\description{
DMC model simulation detailed in  Ulrich, R., Schroeter, H., Leuthold, H., & Birngruber, T. (2015).
Automatic and controlled stimulus processing in conflict tasks: Superimposed diffusion processes and delta functions.
Cognitive Psychology, 78, 148-174. This function uses Base R and is slow. It is recommended to use the function
dmcSim which is a wrapper around the c++ function runDMC.cpp and is much faster.
}
\examples{
\dontrun{
# Example 1
dmc <- dmcSimR()
plot(dmc)

# Example 2
dmc <- dmcSimR(tau = 130)
plot(dmc)

# Example 3
dmc <- dmcSimR(tau = 90)
plot(dmc)

# Example 4
dmc <- dmcSimR(varSP = TRUE)
plot(dmc, "delta")

# Example 5
dmc <- dmcSimR(tau = 130, varDR = TRUE)
plot(dmc, "caf")

# Example 6
dmc <- dmcSimR(stepDelta = 10, stepCAF = 25)
plot(dmc)

}

}
