% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcSims.R
\name{dmcSims}
\alias{dmcSims}
\title{dmcSims}
\usage{
dmcSims(params, printInputArgs = FALSE, printResults = FALSE)
}
\arguments{
\item{params}{(list of parameters to dmcSim)}

\item{printInputArgs}{Print DMC input arguments to console}

\item{printResults}{Print DMC output to console}
}
\value{
list of dmcsim
}
\description{
Run dmcSim with range of input parameters.
}
\examples{
\dontrun{
library(DMCfun)

# Example 1
params <- list(amp = seq(10, 20, 5), tau = c(50, 100, 150))
dmc <- dmcSims(params)
plot(dmc[[1]])          # combination 1
plot(dmc, type = "l")   # delta plots for all combinations

# Example 2
params <- list(amp = seq(10, 20, 5), tau = seq(20, 40, 10), bnds = seq(50, 100, 25))
dmc <- dmcSims(params)
plot(dmc[[1]])  # combination 1
plot(dmc)       # delta plots for all combinations

}

}
