% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcData.R
\name{errDist}
\alias{errDist}
\title{errDist}
\usage{
errDist(n = 10000, proportion = 10)
}
\arguments{
\item{n}{Number}

\item{proportion}{Approximate proportion of errors in percentage}
}
\value{
double
}
\description{
Returns a random vector of 0's (correct) and 1's (incorrect) with
defined proportions (default = 10\% errors).
}
\examples{
# Example 1
x <- errDist(1000, 10)
table(x)

}
