% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcData.R
\docType{data}
\name{flankerData}
\alias{flankerData}
\title{A summarised dataset: see raw data file flankerDataRaw and dmcObservedData.R
This is the summarised Flanker Task data from Ulrich et al. (2015)}
\format{
dmcob
}
\usage{
flankerData
}
\description{
\itemize{
  \item $summary --> Reaction time correct, standard deviation correct, percentage
  error, reaction time incorrect, and standard deviation for incorrect trials
  for both compatible and incompatible trials
  \item $caf --> Proportion correct for compatible and incompatible trials across 5 bins
  \item $delta --> Compatible reactions times, incompatible mean reaction times,
  mean reaction times, incompatible - compatible reaction times (delta), and
  standard deviation + standard error of this difference across 10 bins.
}
}
\keyword{datasets}
