% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flankerData2.R
\docType{data}
\name{flankerData2}
\alias{flankerData2}
\title{A summarised dataset: see raw data file flankerData2.txt and dmcObservedData.R}
\format{dmcob}
\usage{
flankerData2
}
\description{
\itemize{
  \item $summary --> Reaction time correct, standard deviation correct, percentage
  error, reaction tiime incorrect, and standard deviation for incorrect trials
  for both compatible and incompatible trials
  \item $caf --> Proportion correct for compatible and incompatible trials across 5 bins
  \item $delta --> Compatible reactions times, incompatible mean reaction times,
  mean reaction times, incompatible - compatible reaction times (delta), and
  standard deviation + standard error of this difference across 10 bins.
}
}
\keyword{datasets}
