% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcData.R
\docType{data}
\name{flankerDataRaw}
\alias{flankerDataRaw}
\title{Raw flanker data from Ulrich et al. (2015)}
\usage{
flankerDataRaw
}
\description{
\itemize{
  \item Subject Subject number
  \item Comp comp vs. incomp
  \item RT
  \item Error 0 = correct, 1 = error
}
}
\keyword{datasets}
