% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.dmcfit.R
\name{mean.dmcfit}
\alias{mean.dmcfit}
\title{mean.dmcfit}
\usage{
\method{mean}{dmcfit}(x, ...)
}
\arguments{
\item{x}{Output from dmcFit}

\item{...}{pars}
}
\value{
tibble
}
\description{
Mean of the simulation results from dmcFitAgg/dmcFitVPs.
}
\examples{
\dontrun{
# Example 1: Fit individual data then aggregate
fitVPs <- dmcFitVPs(flankerData, nTrl = 1000)
summary(fitVPs)
plot(fitVPs, flankerData, VP = 1)
fitAgg <- mean(fitVPs)
summary(fitAgg)
plot(fitAgg, flankerData)
}

}
