% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcFit.R
\name{mean.dmcfitvp}
\alias{mean.dmcfitvp}
\title{mean.dmcfitvp}
\usage{
\method{mean}{dmcfitvp}(x, ...)
}
\arguments{
\item{x}{Output from dmcFitVPs}

\item{...}{pars}
}
\value{
dmcfit

The function returns a list with the relevant aggregated results dmcFitVPs. The list
is accessed with obj$name and so on with the the following:
\item{obj$means}{means}
\item{obj$delta}{delta}
\item{obj$caf}{caf}
\item{obj$prms}{par}
}
\description{
Aggregated simulation results from dmcFitVPs.
}
\examples{
\donttest{
# Example 1: Fit individual data then aggregate
fitVPs <- dmcFitVPs(flankerData, nTrl = 1000, VP = c(2))
fitAgg <- mean(fitVPs)
plot(fitAgg, flankerData)
}

}
