% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dmcfit_subject}
\alias{plot.dmcfit_subject}
\title{plot.dmcfit_subject: Plot observed + fitted data}
\usage{
\method{plot}{dmcfit_subject}(
  x,
  y,
  subject = NULL,
  figType = "summary",
  labels = c("Compatible", "Incompatible", "Observed", "Predicted"),
  cols = c("green", "red"),
  ylimRt = NULL,
  ylimErr = NULL,
  xlimCDF = NULL,
  ylimCAF = NULL,
  cafBinLabels = FALSE,
  ylimDelta = NULL,
  xlimDelta = NULL,
  xlabs = TRUE,
  ylabs = TRUE,
  xaxts = TRUE,
  yaxts = TRUE,
  xylabPos = 2,
  resetPar = TRUE,
  legend = TRUE,
  legend.parameters = list(legend = c("Observed", "Predicted")),
  ...
)
}
\arguments{
\item{x}{Output from dmcFit}

\item{y}{Observed data}

\item{subject}{NULL (aggregated data across all subjects) or integer for subject number}

\item{figType}{summary, rtCorrect, errorRate, rtErrors, cdf, caf, delta, all}

\item{labels}{Condition labels c("Compatible", "Incompatible", "Observed", "Predicted") default}

\item{cols}{Condition colours c("green", "red") default}

\item{ylimRt}{ylimit for Rt plots}

\item{ylimErr}{ylimit for error rate plots}

\item{xlimCDF}{ylimit for CDF plot}

\item{ylimCAF}{ylimit for CAF plot}

\item{cafBinLabels}{TRUE/FALSE}

\item{ylimDelta}{ylimit for delta plot}

\item{xlimDelta}{xlimit for delta plot}

\item{xlabs}{TRUE/FALSE}

\item{ylabs}{TRUE/FALSE}

\item{xaxts}{TRUE/FALSE}

\item{yaxts}{TRUE/FALSE}

\item{xylabPos}{2}

\item{resetPar}{TRUE/FALSE Reset graphical parameters}

\item{legend}{TRUE/FALSE}

\item{legend.parameters}{list}

\item{...}{additional plot pars}
}
\value{
Plot (no return value)
}
\description{
Plot the simulation results from the output of dmcFit. The plot
can be an overall summary, or individual plots (activation, trials, pdf, cdf,
caf, delta, all). Plot type summary1 contains an activation plot, example
individual trials, the probability distribution function (PDF), the cumulative
distribution function (CDF), the conditional accuracy function (CAF) and
delta plots. This required that dmcSim is run with fullData = TRUE. Plot type
summary2 contains only the PDF, CDF, CAF and delta plots and does not require
that dmcSim is run with fullData = TRUE.
}
\examples{
\donttest{
# Example 1
resTh <- dmcFit(flankerData, nTrl = 5000)
plot(resTh, flankerData)
plot(resTh, flankerData, figType = "deltaErrors")

# Example 2
resTh <- dmcFit(simonData, nTrl = 5000)
plot(resTh, simonData)
}

}
