% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dmcsim.R
\name{plot.dmcsim}
\alias{plot.dmcsim}
\title{plot.dmcsim}
\usage{
\method{plot}{dmcsim}(x, figType = "summary1", newFig = TRUE, ...)
}
\arguments{
\item{x}{Output from dmcSim}

\item{figType}{summary1, summary2, activation, trials, pdf, cdf, caf, delta, all}

\item{newFig}{TRUE/FALSE}

\item{...}{pars}
}
\description{
Plot the simulation results from the output of dmcSim. The plot
can be an overall summary, or individual plots (activation, trials, pdf, cdf,
caf, delta, all). Plot type summary1 contains an activation plot, example
individual trials, the probaility distribution function (PDF), the cumulative
distribution function (CDF), the conditional accuracy functin (CAF) and
delta plot. This requires that dmcSim is run with fullData = TRUE. Plot type
summary2 contains only the PDF, CDF, CAF and delta plots and does not require
that dmcSim is run with fullData = TRUE.
}
\examples{
\dontrun{
library(DMCfun)

# Example 1
dmc = dmcSim(fullData = TRUE)
plot(dmc)

# Example 2
dmc = dmcSim()
plot(dmc)

# Example 3
dmc = dmcSim()
plot(dmc, type = "all")
}

}
