% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randBeta.R
\name{randBeta}
\alias{randBeta}
\title{randBeta}
\usage{
randBeta(n = 1, shape = 3, lim = c(0, 1))
}
\arguments{
\item{n}{The number of random numbers to return.}

\item{shape}{Shape parameter of the underlying beta distribution}

\item{lim}{The min/max limits to which to scale the distribution}
}
\value{
vector
}
\description{
Draw n random numbers from a beta distribution with shape and
scale between limits
}
\examples{
library(DMCfun)

# Example 1
x <- randBeta()

# Example 2
x <- randBeta(10000)
hist(x, 50)

# Example 3
x <- randBeta(100000, 3, lim = c(-75, 75))
hist(x, 50)

# Example 4
x <- randBeta(100000, 1, lim = c(-75, 75))  # uniform with shape = 1
hist(x, 50)

}
