% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcData.R
\docType{data}
\name{simonDataRaw}
\alias{simonDataRaw}
\title{Raw simon data from Ulrich et al. (2015)}
\usage{
simonDataRaw
}
\description{
\itemize{
  \item Subject Subject number
  \item Comp Compatibility condition (comp vs. incomp)
  \item RT Reaction Time
  \item Error Error coding (0 = correct, 1 = error)
}
}
\keyword{datasets}
