% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dmcfit.R
\name{summary.dmcfit}
\alias{summary.dmcfit}
\title{summary.dmcfit}
\usage{
\method{summary}{dmcfit}(object, ...)
}
\arguments{
\item{object}{Output from dmcFitAgg}

\item{...}{pars}
}
\value{
DataFrame (tibble)
}
\description{
Summary of the simulation results from dmcFitAgg
}
\examples{
\dontrun{
# Example 1
fitAgg <- dmcFitAgg(flankerData, nTrl = 1000)
summary(fitAgg)  # this is just fitAgg$means
}

}
