% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dmcfits}
\alias{summary.dmcfits}
\title{summary.dmcfits: dmc fit aggregate summary (2+ data sets)}
\usage{
\method{summary}{dmcfits}(object, digits = 2, ...)
}
\arguments{
\item{object}{Output from dmcFit}

\item{digits}{Number of digits in the output}

\item{...}{pars}
}
\value{
DataFrame
}
\description{
Summary of the simulation results from dmcFit
}
\examples{
\donttest{
# Example 1
fitAggs <- dmcFit(list(flankerData, simonData), nTrl = 1000)
summary(fitAggs)
}

}
