% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dmcfits_subject}
\alias{summary.dmcfits_subject}
\title{summary.dmcfits_subject: dmc fit aggregate summary}
\usage{
\method{summary}{dmcfits_subject}(object, digits = 2, ...)
}
\arguments{
\item{object}{Output from dmcFitAgg}

\item{digits}{Number of digits in the output}

\item{...}{pars}
}
\value{
DataFrame
}
\description{
Summary of the simulation results from dmcFitAgg
}
\examples{
\donttest{
# Example 1
fitsSubject <- dmcFitSubject(list(flankerData, simonData), nTrl = 1000, subjects = c(1:3))
summary(fitsSubject)
}

}
