% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dmcfitvp}
\alias{summary.dmcfitvp}
\title{summary.dmcfitvp}
\usage{
\method{summary}{dmcfitvp}(object, ...)
}
\arguments{
\item{object}{Output from dmcFitVPs}

\item{...}{pars}
}
\value{
list of DataFrames (tibbles) with the first being individual participant fitted parameters and the 
second being the mean fitted parameters
}
\description{
Summary of the simulation results from dmcFitVPs
}
\examples{
\dontrun{
# Example 1
fitVPs <- dmcFitVPs(flankerData, nTrl = 1000, VP = c(1, 10))
summary(fitVPs)
fit <- mean(fitVPs)
}

}
