% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dmcsim}
\alias{summary.dmcsim}
\title{summary.dmcsim: dmc simulation summary}
\usage{
\method{summary}{dmcsim}(object, digits = 1, ...)
}
\arguments{
\item{object}{Output from dmcSim}

\item{digits}{Number of digits in the output}

\item{...}{pars}
}
\value{
DataFrame
}
\description{
Summary of the overall results from dmcSim
}
\examples{
\donttest{
# Example 1
dmc <- dmcSim()
summary(dmc)

# Example 2
dmc <- dmcSim(tau = 90)
summary(dmc)
}

# Example 3
fit <- dmcFit(flankerData, nTrl=100)
summary(fit)

# Example 3
fit <- dmcFitSubject(flankerData, nTrl=100, subjects=c(1,2,3))
summary(fit)

}
