% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcData.R
\name{dmcCombineObservedData}
\alias{dmcCombineObservedData}
\title{dmcCombineObservedData: Combine results from dmcObservedData}
\usage{
dmcCombineObservedData(...)
}
\arguments{
\item{...}{Any number of outputs from dmcObservedData}
}
\value{
dmcobs
}
\description{
Combine observed datasets
}
\examples{
# Example 1
dat <- dmcCombineObservedData(flankerData, simonData)  # combine flanker/simon data
plot(dat, figType = "delta", xlimDelta = c(200, 700), ylimDelta = c(-20, 80),
     cols = c("black", "darkgrey"), pchs = c(1, 2), legend = FALSE, resetPar = FALSE)
legend(200, 80, legend = c("Flanker Task", "Simon Task"),
       col = c("black", "darkgrey"), lty = c(1, 1))

}
