% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcObservedData.R
\name{dmcObservedData}
\alias{dmcObservedData}
\title{dmcObservedData}
\usage{
dmcObservedData(
  dat,
  stepCAF = 20,
  stepDelta = 5,
  outlier = c(200, 1200),
  quantileType = 5
)
}
\arguments{
\item{dat}{Text file containing the observed data or R dataframe (see createDF/addDataDF)}

\item{stepCAF}{Step size for the CAF bins. For example, a step size of 20 would result
in 5 CAF bins centered on 10, 30, 50, 70, and 90\%.}

\item{stepDelta}{Step size for the Delta bins. For example, a step size of 5 would result
in 19 CAF bins positioned at 5, 10, 15, ... 85, 90, 95\%.}

\item{outlier}{Outlier limits e.g., c(200, 1200)}

\item{quantileType}{Argument (1-9) from R function quantile specifying the algorithm (?quantile)}
}
\value{
DataFrame
}
\description{
Basic example analysis script to create data object required
for observed data. Example raw *.txt files are flankerData1.txt, flankerData2.txt,
and simonData1.txt.
}
\examples{
\dontrun{
library(psychReport)
requiredPackages(c("DMCfun",  "ez", "dplyr"))

# Example 1
# datOb <- dmcObservedData("flankerData1.txt")

# Example 2
# datOb <- dmcObservedData("flankerData2.txt")

# Example 3
# datOb <- dmcObservedData("simonData1.txt")

# Example 4
# datOb <- dmcObservedData("simonData2.txt")

#############################################################################
# Example 5 (Basic behavioural analysis from Ulrich et al. 2015)
flankerData1$summaryVP <- tibble::add_column(Task = "flanker", flankerData1$summaryVP, .before = 2)
simonData1$summaryVP   <- tibble::add_column(Task = "simon", simonData1$summaryVP, .before = 2)
datAgg <- rbind(rbind(flankerData1$summaryVP, simonData1$summaryVP))

datAgg$VP   <- factor(datAgg$VP)
datAgg$Task <- factor(datAgg$Task)
datAgg$Comp <- factor(datAgg$Comp)

aovErr <- ezANOVA(datAgg, dv = .(perErr), wid = .(VP), within = .(Comp, Task),
                  detailed = TRUE, return_aov = TRUE)
aovErr <- aovTable(aovErr)
aovDispMeans(aovErr)

aovRt <- ezANOVA(datAgg, dv = .(rtCor), wid = .(VP), within = .(Comp, Task),
                 detailed = TRUE, return_aov = TRUE)
aovRt <- aovTable(aovRt)
aovDispMeans(aovRt)

# Delta
flankerData1$deltaVP <- tibble::add_column(Task = "flanker", flankerData1$deltaVP, .before = 2)
simonData1$deltaVP   <- tibble::add_column(Task = "simon", simonData1$deltaVP, .before = 2)
datDelta             <- rbind(rbind(flankerData1$deltaVP, simonData1$deltaVP))

datDelta <- datDelta \%>\%
  dplyr::select(-c("meanBin", "meanEffect")) \%>\%
  tidyr::gather(Comp, rt, -c("VP", "Task", "bin"))

datDelta$VP   <- factor(datDelta$VP)
datDelta$Task <- factor(datDelta$Task)
datDelta$bin  <- factor(datDelta$bin)
datDelta$Comp <- factor(datDelta$Comp)

aovDelta <- ezANOVA(datDelta, dv = .(rt), wid = .(VP), within = .(Task, Comp, bin),
                    detailed = TRUE, return_aov = TRUE)
aovDelta <- aovTable(aovDelta)

# CAF
flankerData1$cafVP <- tibble::add_column(Task = "flanker", flankerData1$cafVP, .before = 2)
simonData1$cafVP   <- tibble::add_column(Task = "simon", simonData1$cafVP, .before = 2)
datCaf             <- rbind(rbind(flankerData1$cafVP, simonData1$cafVP))

datCaf$VP   <- factor(datCaf$VP)
datCaf$Task <- factor(datCaf$Task)
datCaf$bin  <- factor(datCaf$bin)
datCaf$Comp <- factor(datCaf$Comp)

aovCaf <- ezANOVA(datCaf, dv = .(accPer), wid = .(VP), within = .(Task, Comp, bin),
                  detailed = TRUE, return_aov = TRUE)
aovCaf <- aovTable(aovCaf)
aovDispMeans(aovCaf)

}
# Example 6
dat <- createDF(nVP = 50, nTrl = 50,
                design = list("Comp" = c("comp", "incomp")))
dat <- addDataDF(dat,
                 RT = list(list(c("Comp:comp"), vals = c(500, 75, 120)),
                           list(c("Comp:incomp"), vals = c(530, 75, 100))),
                 Error = list(list(c("Comp:comp"), vals = c(3, 2, 2, 1, 1)),
                            list(c("Comp:incomp"), vals = c(15, 3, 2, 1, 1))))
datOb <- dmcObservedData(dat)
plot(datOb)
plot(datOb, VP = 1)

}
