% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dmcfit.R
\name{plot.dmcfit}
\alias{plot.dmcfit}
\title{plot.dmcfit}
\usage{
\method{plot}{dmcfit}(x, y, figType = "summary", newFig = TRUE,
  VP = NULL, ...)
}
\arguments{
\item{x}{Output from dmcFit}

\item{y}{Observed data}

\item{figType}{summary, rtCorrect, errorRate, rtErrors, cdf, caf, delta, all}

\item{newFig}{TRUE/FALSE}

\item{VP}{NULL (aggregated data across all participants) or integer for participant number}

\item{...}{pars}
}
\description{
Plot the simulation results from the output of dmcSim. The plot
can be an overall summary, or individual plots (activation, trials, pdf, cdf,
caf, delta, all). Plot type summary1 contains an activation plot, example
individual trials, the probaility distribution function (PDF), the cumulative
distribution function (CDF), the conditional accuracy functin (CAF) and
delta plots. This required that dmcSim is run with fullData = TRUE. Plot type
summary2 contains only the PDF, CDF, CAF and delta plots and does not require
that dmcSim is run with fullData = TRUE.
}
\examples{
\dontrun{
library(DMCfun)

# Example 1
resTh <- dmcFitAgg(flankerData1, nTrl = 5000)
plot(resTh, flankerData1)

# Example 2
resTh <- dmcFitAgg(flankerData2, nTrl = 5000)
plot(resTh, flankerData2)
plot(resTh, flankerData2, figType = "all")

# Example 3
resTh <- dmcFitAgg(simonData1, nTrl = 5000)
plot(resTh, simonData1)

# Example 4
resTh <- dmcFitAgg(simonData1, nTrl = 5000, VP = 1)
plot(simonData1, VP = 1)
plot(resTh, simonData1, VP = 1)
}

}
