% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dmcsim.R
\name{plot.dmcsim}
\alias{plot.dmcsim}
\title{plot.dmcsim}
\usage{
\method{plot}{dmcsim}(
  x,
  figType = "summary1",
  newFig = TRUE,
  ylimCAF = c(0, 1),
  cafBinLabels = FALSE,
  ylimDelta = c(-50, 150),
  ylimRt = c(200, 800),
  ylimErr = c(0, 20),
  errorBars = TRUE,
  ...
)
}
\arguments{
\item{x}{Output from dmcSim}

\item{figType}{summary1, summary2, summary3, activation, trials, pdf, cdf, caf, delta, rtCorrect, rtErrors, errorRate, all}

\item{newFig}{TRUE/FALSE}

\item{ylimCAF}{ylimit for CAF plot}

\item{cafBinLabels}{TRUE/FALSE}

\item{ylimDelta}{ylimit for delta plot}

\item{ylimRt}{ylimit for rt plot}

\item{ylimErr}{ylimit for er plot}

\item{errorBars}{TRUE/FALSE}

\item{...}{additional plot pars}
}
\description{
Plot the simulation results from the output of dmcSim. The plot
can be an overall summary, or individual plots (activation, trials, pdf, cdf,
caf, delta, all). Plot type summary1 contains an activation plot, example
individual trials, the probability distribution function (PDF), the cumulative
distribution function (CDF), the conditional accuracy function (CAF) and
delta plot. This requires that dmcSim is run with fullData = TRUE. Plot type
summary2 contains only the PDF, CDF, CAF and delta plots and does not require
that dmcSim is run with fullData = TRUE.
}
\examples{
\dontrun{
library(DMCfun)

# Example 1
dmc = dmcSim(fullData = TRUE, stepCAF = 10)
plot(dmc)

# Example 2
dmc = dmcSim()
plot(dmc)

# Example 3
dmc = dmcSim()
plot(dmc, figType = "all")

# Example 4
dmc = dmcSim()
plot(dmc, figType = "summary3")
}

}
