% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dmcfit.R
\name{summary.dmcfit}
\alias{summary.dmcfit}
\title{summary.dmcfit}
\usage{
\method{summary}{dmcfit}(object, ...)
}
\arguments{
\item{object}{Output from dmcFit}

\item{...}{pars}
}
\value{
tibble
}
\description{
Summary of the simulation results from dmcFitAgg/dmcFitVPs.
}
\examples{
\dontrun{
library(DMCfun)

# Example 1
fitAgg <- dmcFitAgg(flankerData, nTrl = 1000)
summary(fitAgg)
plot(fitAgg, flankerData)

# Example 2
fitVPs <- dmcFitVPs(flankerData, nTrl = 1000, VP = c(1, 10))
summary(fitVPs)
fit <- mean(fitVPs)
plot(fit, flankerData)
}

}
