% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dmcfit}
\alias{summary.dmcfit}
\title{summary.dmcfit}
\usage{
\method{summary}{dmcfit}(object, digits = 2, ...)
}
\arguments{
\item{object}{Output from dmcFitAgg}

\item{digits}{Number of digits in the output}

\item{...}{pars}
}
\value{
DataFrame
}
\description{
Summary of the simulation results from dmcFitAgg
}
\examples{
\donttest{
# Example 1
fitAgg <- dmcFitAgg(flankerData, nTrl = 1000)
summary(fitAgg)  
}

}
