% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dmcsim.R
\name{summary.dmcsim}
\alias{summary.dmcsim}
\title{summary.dmcsim}
\usage{
\method{summary}{dmcsim}(object, ...)
}
\arguments{
\item{object}{Output from dmcSim}

\item{...}{pars}
}
\value{
list
}
\description{
Summary of the simulation results from dmcSim
}
\examples{
\dontrun{
# Example 1
dmc <- dmcSim()
summary(dmc)
}

}
